publisher_service/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.66.0
h11==0.12.0
idna==2.10
importlib-metadata==4.6.1
pika==1.2.0
pydantic==1.8.2
PyMySQL==1.0.2
python-dotenv==0.18.0
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
six==1.16.0
starlette==0.14.2
SQLAlchemy==1.3.23
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
zipp==3.5.0
